Added "No Oppositions" display to Listing Page

+++ application/views/admin/opposition/list.php
@@ -1,6 +1,5 @@
 <?php
-echo $pagination;
-if (count($oppositions) > 0) { ?>
+echo $pagination; ?>
     <table>
         <thead>
             <tr>
@@ -11,16 +10,22 @@ if (count($oppositions) > 0) { ?>
         </thead>
         <tbody>
     <?php
-    foreach ($oppositions as $opposition) { ?>
+    if (count($oppositions) > 0) {
+        foreach ($oppositions as $opposition) { ?>
             <tr>
                 <td><?php echo $opposition->name; ?></td>
                 <td><a href="/admin/opposition/edit/id/<?php echo $opposition->id;?>"><?php echo $this->lang->line('opposition_edit'); ?></a></td>
                 <td><a href="/admin/opposition/delete/id/<?php echo $opposition->id;?>"><?php echo $this->lang->line('opposition_delete'); ?></a></td>
             </tr>
     <?php
+        }
+    } else { ?>
+            <tr>
+                <td colspan="3"><?php echo $this->lang->line('opposition_no_oppositions'); ?></td>
+            </tr>
+    <?php
     } ?>
         </tbody>
     </table>
 <?php
-}
 echo $pagination; ?>
